Fix crash when Pixelate is applied on Depth pass
The node tried to be smart w.r.t the socket type.

While with the previous simple pass-through of rounded coordinate
did not require any special code for such flexibility, now it will
be more tricky to support it.

There are likely to be other sources of slowdown and memory usage,
and the node is likely to be used on a color anyway, so choose simpler
way of dealing with the data type.