VSE: bilinear upscaling no longer adds transparent border around the image
The code is actually within BLI, but it is mostly used by VSE (and
somewhat by Compositor). Ever since low level "bilinear/bicubic"
interpolation functions were added back in 2007 (9298357670f5), they
had different behavior in how a pixel that is right on the edge
of an image is treated: bilinear was blending in transparent black,
whereas bicubic was clamping samples to the image edge.

Make bilinear do the same as bicubic does, which is clamping samples
to the edge of an image. For typical use cases this better matches
user expectations: they put in low resolution image into say
sequencer, it is upscaled, and the whole image covers the whole output,
without any transparency appearing aroud the edges.

Performance of bilinear filter (measuring IMB_transform on 4K image)
seems to be unchanged.