Fix Split node in full-frame compositor when no inputs are connected
Not having inputs connected triggers the constant folding code paths
in the operation builder, which attempts to convert the operation to
a constant memory buffer. This happens before the canvas of the node
is known, and the expected result in area of 1x1 pixels.

The split node can not be constant-folded in such way without causing
regression compared to the tiled compositor: for unconnected inputs
the node is expected to use the output resolution as canvas.

IN theory the node can be constant folded if the split percentage is
0 or 1, but there is no mechanism which allows such option-dependent
constant folding check.

This fixes node_split_unlinked.blend regression test which fails when
full-frame compositor is used by default.