Vulkan: Refactor: Specialization Constants
VkSpecializationInfo and VkSpecializationMapEntry should only be used immediately before generation,
as they can cause confusion during multi-threaded execution. Unique_ptr is also used for the same reason.
The question is, can we do anything with Shader-Bind?
At this point, it is impossible to generate a Pipeline, so if we do not make it a through path, it will be completely broken.
I tried doing a simple Shading of EEvEE_Next on another branch.
Setting the constants.is_dirty flag to false in shader-bind makes pipeline generation probabilistic.