Utilize depth input in the preprocess stage.
This adds two new variations for the functions: 'attribute_foreach' and 'gather_attributes_for_propagation'. The functions are the same but also accept a VArray containing the depth value for each instance component.

There is still a 'known bug' that a fully realized component (such as a mesh, etc.) may get attributes from a realized instance that doesn't hold any mesh component itself. This behavior may be debatable in some scenarios.

After some thought, I decided that the current design is okay-ish.