UI: Tree view, make expand-on-click optional
Make the 'expand-on-click' behavior of tree items optional. It can now
be determined per implementation.

Instead of calling `tree_item.set_collapsed(false)` directly, it is
replaced with a call to `tree_item.on_click()`. The base implementation
of that function still calls `set_collapsed(false)`; since this is a
virtual function, it can be overridden by subclasses.

---------------

This was the least-intrusive change I could think of. I'm not a huge fan
of this approach, though, as it makes the overridden implementation
dependent on the exact implementation of the base function.

Personally I think having the base function empty would be better; in
that case any subclass that wants to retain the expand-on-click behavior
should call `set_collapsed(false)` in their override.

Alternatively we could have some flag that controls this behavior. That
flag can then be set on subclasses, or even controlled by a user
preference. Not sure if that's the way we want to go with this, though.