Fix slow toggling of scene tool settings
For example, snapping or pivot point.

The slowdown happens in complex files with drivers which use scene
properties as input. Can be seen in the demo file from #117335.

This change adds ability to mark RNA property as required to bring
the evaluated ID up-to-date, but not do possibly expensive tag for
drivers.

The tool settings possibly can be tagged as do-not-do-deg-tag at
all, but safer seems to follow the current behavior a bit closer.