Cleanup: OpenGL program creation and linking
This PR sanitizes the place where glPrograms are created. Previously it
was done when the first shader stage was compiled. when creating a shader
stage it was directly bound to the program. This creates a weird control
flow.

This PR will generate and attach the shader stages inside link program.
Ensuring that create_shader_stage doesn't alter the program, which isn't
clear in its name.