Compositor: Only allow Output nodes in root node tree
This patch makes it such that output nodes like the Composite and File
Output nodes are only considered if they are in the root node tree, that
is, not in node group.

There are a number of reasons why this shouldn't be allowed:

- The active output tagging mechanism only works in the same node tree,
  not across trees. So all output nodes will be active at the same time,
  fighting over which one will write the final output, where the last
  added node or node group will most likely be the winner.
- Node groups are shareable resources, and it wouldn't be safe for the
  user to append a node group and unknowingly get files written to disk.
  Therefore, we should only allow outputs that the user knowingly added
  explicit to the compositor node tree.
- Inexperienced users might not know that their output comes from inside
  a node group, which can lead to confusion.

The GPU compositor already implements this mechanism, so we only adjust
the CPU implementation, as well as the verification mechanism in the
render pipeline code.