EEVEE: Hiz Allocation Size
EEVEE uses an HiZ buffer. The size of the HiZ buffer is based on the render extent.
Probes assume that the render extent is always bigger than needed for rendering
probes. This assumption is incorrect. Although this PR fixes
this by allocating the
required size, it is still renders with artifacts. These artifacts originate from the
lighting module and also needs to be fixed.