RenderTests: Fix Incorrect Command in Report
When running the render test for EEVEE-Next the command printed on
the report to update the reference images are incorrect.

The cause of this is that the title of the report is used to create
the command. EEVEE-Next uses a space in their title which generates
an incorrect command.

This PR is a quick fix which replaces spaces with underscores.