Fix #117076: 'Selection to Cursor' ignores Object Constraints
Use the implementation in the transform system that already considers
Constraints.

---
To support this, one solution would be to expose the internal transform
function `bool constraints_list_needinv(TransInfo *t, ListBase *list)`
and use `pchan->constinv`.

But with each bug report about Snap Selection operators, I notice that
the expected behavior is to repeat the same behavior as the Move
operator (which supports Crazy Space, Constraints and other features).

Therefore, the solution proposed here is to create the
`ED_transform_move_to_location` function that reuses the transform
system implementation in the snap operator.