Compositor: Rewrite and optimize Double Edge Mask node
This patch rewrites and optimizes the Double Edge Mask node to be orders
of magnitude faster. For a 1k complex mask, it is 650x faster, while for
a 1k simple mask, it is only 50x faster.

This improvement is attributed to the use of a new Jump Flooding
algorithm as well as multi-threading, matching the GPU implementation.

The result of the new implementation differs if the Inner Edge option is
set to Adjacent Only and the Inner Mask is not inside the Outer Mask.
The difference is due the condition that identifies the edges of the
Inner Mask.