Fix #117539: Image.scale() always works on the first tile
`BKE_image_scale` -- which is only used for the python API -- was
getting the `ImBuf` without providing an `ImageUser`.
This is fine, but always gets the first tile.

In order to respect `Image.tiles.active_index` for the scaling
operation, we can create a default `ImageUser` and set its `tile`.

Afaict, no other members of the `ImageUser` need to be set for tiled
images, so should be no other functional changes (compared to passing a
nullptr).