Addon: catch certain ZIP packaging errors in addon_install
One error that occurs when packaging a multi-file addon into a
ZIP is zipping the contents of the addon folder, instead of the
addon folder itself. When installing the ZIP using addon_install,
the files that should be inside a folder instead get extracted
into the top-level of the script directory. There is also no user
feedback about what went wrong.

addon_install can detect this error by the presence of an
__init__.py file at the top-level of the ZIP archive, and report
an error. AFAIK there is no use-case for having an __init__.py
in the top-level of the script directory, so there are no false
positives.