UI: Tooltip when bone collections are linked
Linked bone collections are disabled in b79419914a34023f0644df1091803103956df64c
Include 'Disabled' tooltip for `move and assign to collection`
operation when bcol are linked.
In poll functions all other cases are already handled and also not
possible to get the index of hovered collection so just include
`poll_msg_set` in the end.