Fixes setting object bound dimensions using Multiple Value Edition from UI. Was failing, according to message set on RNA info due to cached values.
Solves it resetting the cached bounds, and updating the local
object_to_world matrix.

It does not recompute the min-max values based on geometry as this
stored on another cache and is not necessary because current
implementations works changing the scale vector of the object. The
matrix must be set to match the scale, as the value is recovered just
after set.

It also adds the values on Object Panel.