Nodes: expose socket location to the Python API
When writing Python scripts that change a node tree layout, it is
useful to have a way to get where the sockets are located in the node
tree*.

One drawback of this is that sockets' locations are stored in the
socket `runtime` structure, and are only computed when the node tree
is actually drawn on screen. This results in a property that is
useless in background mode, or when creating node trees in a
non-interactive context. This issue is documented in the property
description.

* One example is the Node Wrangler add-on, which includes a tool to
add reroute nodes next to a given node's outputs. It currently uses an
approximation, which typically results in misaligned reroutes.
Another example is the third-party Voronoi Linker add-on, which goes
so far as to use ctypes to access the C structure for the socket
runtime.