Geometry Nodes: Add local/global toggle to Rotate Rotation node
All it does internally is reverse the order of the rotations, but this
is the same as old "Rotate Euler" node, and it's a bit more intuitive
this way. Also use the same socket names, "Rotation" and "Rotate
By' which are much more intuitive.

Technically it would be better to not have to duplicate the multi-
function just to switch the order of the arguments. But the
evaluator assumes that the order always matches the socket
order currently.