Outliner: Add shortcut to Add Object and Duplicate
Object operations like Move Collection are already available in
the Outliner, but basic operations like Add or Duplicate, are not.

Add shortcuts to the Outliner for consistency and convenience:
* Add Object (Shift+A)
* Duplicate (Shift+D)
* Duplicate Linked (Alt+D)

For the Industry Compatible keymap, the Duplicate shortcut is Ctrl+D.

There is one conflict with the Expand/Collapse toggle, which was using
Shift+A. I've removed the shortcut (which can be added back manually
from the Context > View menu), and added the operator to the View pie
for quick access. Worth mentioning that expanding/collapsing can be
done as well by shift-clicking on the arrow next to the item.