RenderTest: Fix EEVEE Render Test
Panorama dicing test fails for EEVEE. EEVEE creates a shader interface
that isn't compatible with the vulkan backend. This PR disables the check.

Check should be enabled again after EEVEE has been replaced by EEVEE-Next.
This PR also changes the behvior when checks are executed. It used to be
executed when blender was build with asserts. Now it is behind the --debug-gpu flag.