Curves: Control the radius of each individual hair strand
Before it was not possible to control the radius of each curve.
The variables defined in the shader: hairRadRoot, hairRadTip and hairRadShape controlled the appeareance of all curves.

I added a new buffer that has for each hair strand a radiusBottom, radiusTop and radiusShape. The buffer is currently only used by the `draw_cuves.cc` and not by the old hair system.
In order to not break the old hair system, I added a toggle in the shader called `usePerHairStrandRadius` that switches between drawing using a custom radius for each hair strand and using
a global radius configuration. The old hair system still uses a single radius configuration.