Vulkan: Refactor: VkPipeline
This change is a result of render pass rationalization.
Limit VkRenderpass generation to a minimum number of times.
Make the viewport and scissors dynamic.
These two refactorings significantly reduce the number of VkPipeline generations.
The pipeline cache implementation is a rough sketch.
In practice, very few pipelines need to cache VertexInput.
First, I'd like to run WorkBench and verify this.